Time: 0 ms (100.00%), Space: 6.5 MB (87.75%) - LeetHub

+++ 818-race-car/818-race-car.cpp
@@ -11,7 +11,7 @@ public:
                 pair<int, int> node = q.front();
                 q.pop();
                 int pos = node.first;
-                long int vel = node.second;
+                int vel = node.second;
                 if (pos == target) {
                     return depth;
                 }
